@@ -10,8 +10,8 @@
         }
     ],
     "require": {
-        "php": ">=5.3.9",
-        "kigkonsult/icalcreator": "2.22.*",
+        "php": ">=5.4",
+        "kigkonsult/icalcreator": "2.24.*",
         "swiftmailer/swiftmailer": ">=4.2.0,~5.0",
         "symfony/config": "~2.3|~3.1",
         "symfony/dependency-injection": "~2.3|~3.1",
